Add HelpMojo generation
This adds support for generating a Maven HelpMojo. A new task for generating the
HelpMojo sources is added and wired to that the generated class gets
compiled. After MojoDescriptor extraction and additional step for moving
the HelpMojo to the correct package is executed. This rewrites the
source and class files to place the HelpMojo in the package where the
most user implemented mojos have been found.