Drop functionality to extract mojos from dependency projects
Extracting data from one project in another is not the way this use case
should be solved in Gradle. The idiomatic way of doing this is to apply
this plugin to each project that generates a mojo descriptor and then
expose a variant for the mojo descriptor that can be consumed by an
aggregator project. That aggregator would then merge the mojo
descriptors together when building a fat jar.