Fix failure in DeprecationsFuncTest
This test is currently failing because the build configures a Java 8 toolchain.
When the test is executed, the test VM is started with Java 8 and the forked
Gradle process is also started with Java 8. This results in a deprecations because
starting with Gradle 9.0 it requires at least Java 17. To work around this, the
project is now compiled with a Java 8 toolchain, but tests execute on Java 21.