Fix failure in DeprecationsFuncTest
This test is currently failing because the build configures a Java 8 toolchain.
When the test is executed, the test VM is started with Java 8 and the forked
Gradle process is also started with Java 8. This results in a deprecation because
starting with Gradle 9.0 it requires at least Java 17. To work around this, the
project is now compiled with a Java 8 toolchain, but tests execute on Java 17.

Unfortunately this caused a new problem with the cross version tests:
Gradle only started to support running on Java 17 with the 7.3 release.
For that reason compatibility of the plugin was raised from 7.0.2 to 7.3
because supporting older versions would complicate the testing setup too
much. Older version might still work but are not tested against.