Fix cross-project access deprecation
This replaces the code to extract UpstreamProjectDescriptor with a
solution that relies on dependency management features.

First, in order to get upstream project classes directories, it uses
an artifact view filtering for project dependencies in the compile classpath.
Then, in order to get upstream project sources directories, it usees an
artifact view to filter for project dependencies and selects a sources
variant from the artifacts in the compile classpath.
The last step is to group sources and classes together and feed them
into the generator task.

Fixes #283